+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added new SSL certificate of [parley.nu](https://parley.nu) for SSL pinning.
 - Fixed a crash when sending or receiving a new message when there is no welcome message.
 - Fixed a crash when setting the user information after calling `Parley.configure()`.
+- Fixed a crash that could happen due to keyboard visiblity.
 - When setting or clearing the user information, Parley will now reconfigure itself to show the contents of the corresponding chat.
 - Calling `Parley.configure()` twice without resetting Parley is unsupported and will now throw an error when this happens.